Code for CSV & Flexible Comparisons
This first commit contains a basic working version of the project as it
has been invisioned. It contains a base class that encapsulates basic
functionality for working with arrays of dictionaries, including match
and diff comparisons. The class contains functionality for taking the
data in each dictionary in the array and re-mapping them by specifying a
new key made up of a tuple of values from each dictionary in the array.
This is the key concept of this project.

This commit features 2 child objects--one for creating instances from
CSV files (LookupMap) and another (Result), which is used to store the
results of comapring 2 LookupMap objects. These two objects make it
possible to compare LookupMap objects with eachother, save the results,
and then perform further calculations.

More functionality for database query results is likely to be added
soon.