Clean Up of Revised Script
This adds some basic features that clean up the very rudimentary code
that was added in the first commit to make it more Pythonic. First, the
script ran a test when it ran, which has now been wrapped in a __name__
== "__main__" statment. As well, there is now a blank __init__ file to
make the project into a basic pacakge for the time being. This will
evlolve as a package later, and the revised_lookup module will
eventually be renamed.