Modifications to Write Method
This changes the write method (which I never bothered to get working as
it should) so that the user can no longer specify args that represent
fields that are to be written to the output csv file. That functionality
should be added in later, but for now, the method is simplified and
works as it should.